feat(api): unify observations endpoint
